update unit tests - assert widget placed on canvas and update snaphots
